last-update

+++ app/api/events/[id]/route.ts
@@ -1,26 +1,14 @@
 import { NextResponse } from 'next/server'
-import { createRouteHandlerClient } from '@supabase/auth-helpers-nextjs'
-import { cookies } from 'next/headers'
 import { Database } from '@/lib/supabase-types'
+import { getSessionAndRole } from '@/lib/checkrole'
 
 export async function PUT(request: Request, { params }: { params: { id: string } }) {
-  const supabase = createRouteHandlerClient<Database>({ cookies })
-  const body = await request.json()
-
-  const {
-    data: { session }
-  } = await supabase.auth.getSession()
-
-  if (!session) {
-    return NextResponse.json({ error: 'Yetkisiz erişim' }, { status: 401 })
+  const { error, session, role, supabase } = await getSessionAndRole()
+  if (error || !session) {
+    return NextResponse.json({ error }, { status: 401 })
   }
 
-  const { data: profile } = await supabase
-    .from('profiles')
-    .select('role')
-    .eq('id', session.user.id)
-    .single()
-
+  const body = await request.json()
   const updates: Database['public']['Tables']['events']['Update'] = {}
 
   if ('title' in body) updates.title = body.title
@@ -35,46 +23,35 @@ export async function PUT(request: Request, { params }: { params: { id: string }
 
   let query = supabase.from('events').update(updates).eq('id', params.id)
 
-  if (profile?.role !== 'admin') {
+  if (role !== 'admin') {
     query = query.eq('user_id', session.user.id)
   }
 
-  const { data, error } = await query.select('*').single()
+  const { data, error: updateError } = await query.select('*').single()
 
-  if (error) {
-    return NextResponse.json({ error: error.message }, { status: 500 })
+  if (updateError) {
+    return NextResponse.json({ error: updateError.message }, { status: 500 })
   }
 
   return NextResponse.json(data)
 }
 
 export async function DELETE(_: Request, { params }: { params: { id: string } }) {
-  const supabase = createRouteHandlerClient<Database>({ cookies })
-
-  const {
-    data: { session }
-  } = await supabase.auth.getSession()
-
-  if (!session) {
-    return NextResponse.json({ error: 'Yetkisiz erişim' }, { status: 401 })
+  const { error, session, role, supabase } = await getSessionAndRole()
+  if (error || !session) {
+    return NextResponse.json({ error }, { status: 401 })
   }
-
-  const { data: profile } = await supabase
-    .from('profiles')
-    .select('role')
-    .eq('id', session.user.id)
-    .single()
 
   let query = supabase.from('events').delete().eq('id', params.id)
 
-  if (profile?.role !== 'admin') {
+  if (role !== 'admin') {
     query = query.eq('user_id', session.user.id)
   }
 
-  const { error } = await query
+  const { error: deleteError } = await query
 
-  if (error) {
-    return NextResponse.json({ error: error.message }, { status: 500 })
+  if (deleteError) {
+    return NextResponse.json({ error: deleteError.message }, { status: 500 })
   }
 
   return NextResponse.json({ success: true })
